Split out qSupported tests
