+++ lldb/test/API/tools/lldb-server/supported/Makefile
@@ -1,0 +1,3 @@
+C_SOURCES := main.c
+
+include Makefile.rules
