+++ lldb/test/API/tools/lldb-server/supported/main.c
@@ -1,0 +1,1 @@
+int main() { return 0; }
